feat: add patch implementation

+++ migrations/20251002120231_create_tasks_table.sql
@@ -7,7 +7,7 @@ CREATE TABLE tasks (
     description TEXT,               -- optional
     status VARCHAR(20) NOT NULL,    -- e.g. "pending", "in_progress", "done"
     due_date TIMESTAMPTZ NOT NULL,  -- stores date/time in UTC
-    user_id INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    user_id BIGSERIAL NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
